feat: add relations between User & Record models

+++ routes/modules/records.js
@@ -10,9 +10,10 @@ router.get('/new', async (req, res) => {
 
 router.get('/:id/edit', async (req, res) => {
   try {
-    const id = req.params.id
+    const userId = req.user._id
+    const _id = req.params.id
     const categories = await Category.find().lean()
-    const record = await Record.findById(id).lean()
+    const record = await Record.findOne({ _id, userId }).lean()
     categories.map(category => {
       return category.isSelected = category._id.toString() === record.categoryId.toString() // 新增屬性讓前端知道此筆資料原來的類別
     })
@@ -25,10 +26,11 @@ router.get('/:id/edit', async (req, res) => {
 
 router.put('/:id', async (req, res) => {
   try {
+    const userId = req.user._id
     const _id = req.params.id
     const { name, date, category, amount } = req.body
     const matchCategory = await Category.findOne({ name: category })
-    await Record.findOneAndUpdate({ _id },
+    await Record.findOneAndUpdate({ _id, userId },
       { name, date, amount, categoryId: matchCategory._id })
     res.redirect('/')
   } catch (err) {
@@ -38,8 +40,9 @@ router.put('/:id', async (req, res) => {
 
 router.delete('/:id', async (req, res) => {
   try {
+    const userId = req.user._id
     const _id = req.params.id
-    await Record.findOneAndDelete({ _id })
+    await Record.findOneAndDelete({ _id, userId })
     res.redirect('/')
   } catch (err) {
     console.log(err)
@@ -48,12 +51,13 @@ router.delete('/:id', async (req, res) => {
 
 router.post('/', async (req, res) => {
   try {
+    const userId = req.user._id
     const { name, date, category, amount } = req.body
     const categories = await Category.find().lean()
     const matchCategory = categories.find(c => {
       return c.name.toString() === category
     })
-    await Record.create({ name, date, categoryId: matchCategory._id, amount })
+    await Record.create({ name, date, categoryId: matchCategory._id, amount, userId })
     res.redirect('/')
   } catch (err) {
     console.log(err)
